Add clarrifying comment to poly_chknorm
It is a little unintuitive why in poly_chknorm it suffices to check that
the coefficients of absolute value less than B even though the inputs are NOT
canonically reduced.

The answer is:
* Specification: The definition of the infinity norm in FIPS-204 requires
*                signed canonical reduction prior to applying the bounds check.
*                However, `-B < (a mod± MLDSA_Q) < B` is equivalent to
*                `-B < a < B` under the assumption that
*                `B <= MLDSA_Q - REDUCE32_RANGE_MAX` (cf. the assertion in
*                the code). Hence, the present spec and implementation are
*                correct without reduction.

Before we were using B <= (MLDSA_Q - 1) / 8) (as it is checked in the
reference implementation), but the more natural bound is the slightly
larger MLDSA_Q - REDUCE32_RANGE_MAX.

This commits changes the pre-condition, adds a comment to poly_chknorm, and
also adds an assertion for what I said above. CBMC proves this just fine.

Resolves https://github.com/pq-code-package/mldsa-native/issues/504

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>