HashML-DSA: Add specialized shake256 pre-hashing function
The previous commit added the _pre_hash_internal functions that expect the
pre-hashed message, i.e., requiring the callee to compute the pre-hash
themselves.
The alternative is to perform the pre-hashing inside mldsa-native which eases
the use for the consumer. However, that would require maintaining 12 hash
functions and allowing consumers to replace those with their own hash
functions.
This commit implements a compromise: We do add an API that includes the
pre-hashing step, but only for the most popular pre-hashing hash function:
SHAKE256.
If later a consumer also requires other hash functions, we can add it.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>