Use consistent syntax for macro definitions and invocations
Most assemblers seems to support .macro definitions/invocations with and
without commas between parameters. Some, however, require a comma. In
the code base, we mix the two styles on the codebase -- sometimes even
in the same file.

This commit adjusts autogen to automatically adjust definitions and
invocations of assembly macros to use comma-separated arguments.

Signed-off-by: Hanno Becker <beckphan@amazon.co.uk>