AArch64: Avoid overread in polyz_unpack
Similar as in the previous commit adjusting the AVX2 code, there is a similar
overread in the AArch64 assembly (here by 12 bytes for polyz_unpack_17_asm,
and 8 bytes for polyz_unpack_19_asm).
This commit fixes that by loading the appropriate number of bytes which
is much more elegant than on x86_64.

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>