x86_64: Autogenerate the entire constant array qdata for NTT/INTT
The constant array required for the x86_64 consists of multiple parts:
first some constants like q and q^-1 and then the twiddle factors required
for the NTT.
Thus far, the autogeneration script would only generate the twiddle factors,
while the other constants were hardcoded in consts.c.
However, for HOL-Light we require to generate the entire constant array -
this was introduced in https://github.com/pq-code-package/mldsa-native/pull/811.
Having two diffent functions generating the constants is a risk that they may
eventually diverge.
This commit fixes this by autogenerating the entire qdata in consts.c
(the x86_64_zetas.i is no longer needed). This way we can use the same
generation function for both consts.c and the HOL-Light proofs.

While at it, we also autogenerate the offsets in consts.h

Signed-off-by: Matthias J. Kannwischer <matthias@kannwischer.eu>